move minja to common/minja
