+++ common/chat.cpp
@@ -1,9 +1,9 @@
 #include "chat.hpp"
 #include <json.hpp>
-#include "chat-template.hpp"
 #include "json-schema-to-grammar.h"
 #include "log.h"
-#include "minja.hpp"
+#include "minja/chat-template.hpp"
+#include "minja/minja.hpp"
 
 namespace minja {
     class chat_template;
